requirements.txt
pdfminer.six
PyPDF2
llama-cpp-python
selenium
